Updated and tested the session creation process
prepross.py: I added a new way of updating the 'new_data' dataframe. Also,I deleted
the tqdm function from for loops to make the output cleaner. I completely tested the fucntions and they work fine. In a nutshell, this module first load the data according to the config. Then, it can detect the language of the choosen columns. Finally, it will detect the session and write an id for every session. It uses a time threshold defined in config to create sessions. The id of each session is 'U#i#S#j#', where #i# stands for user id and #j# denotes the session number for user i.

utils.py: it contains the Config class that contains the main configurations needed for data preprocessing. Also it contains the Language Identification class, which I commented as I did not have the fasttext library installed.

prototype.ipynb: it is an ipython notebook, which was used for testing each function.