Added three modules
get_vocab: this module prepares a vocabulary object for the givven data and pickle it to the experiment directory with the name vocab.plk

prepare_train_data: prepares train data from the givven dataset. At the end data has the following header: user_id len_in input output. input and output are sequences of tokens

train_valid_test_split: it splits the givven data to train validation and test sets. Also, it deletes items from valid and test set that are not in the validation set.